+++ extensions/selection-workbench/main.js
@@ -1,7 +1,7 @@
 app.extension({
   id: 'examples.selection-workbench',
   name: 'Selection Workbench',
-  version: '1.2.1'
+  version: '1.3.0'
 });
 
 var help = app.assets.json('data/help.json');